backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
google-generativeai==0.3.2
